Connected FrontEnd to Backend
Added the need code to connect to back end and fornt end.  Created the modals for Adding vaccines and adding doses. They were made into components and imprted them into the Widget component. Moved the .env file into the backend root so that it can load environment variables properly.  Made it so that the server connects to the database as soon as the server is up. Then closes it when the server goes down.